Story
Please write a program which prints out the following story. The user gives a name and a year, which should be inserted into the printout.

Sample output
Please type in a name: Mary
Please type in a year: 1572

Mary is a valiant knight, born in the year 1572. One morning Mary woke up to an awful racket: a dragon was approaching the village. Only Mary could save the village's residents.