Seven Brothers
Please write a function named seven_brothers. When the function is called, it should print out the names of the seven brothers in alphabetical order, as in the example below. See the similarly named exercise in part 1 for more details about the brothers.

Sample output
Aapo
Eero
Juhani
Lauri
Simeoni
Timo
Tuomas